+++ CHANGELOG.md
@@ -4,11 +4,29 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+- N/A
+
+## [1.3.2]
+
+- Fixed bug that caused Yahoo leagues that had unfinished matchups to raise an exception
+
+## [1.3.1]
+
+- Fixed bug where Yahoo consolation matchups could count as a championship matchup
+
+## [1.3.0]
+
+- Updated versions of dependent libraries
+- Added ability to find multiple years for Yahoo leagues
+
+## [1.2.0]
+
 - Added better type checking for validation
+- Fixed bug in YahooLeagueLoader that caused tied matchups to raise an exception
 
 ## [1.1.1]
 
-- [ESPN League Loader] Added check for teams on bye week.
+- For ESPN leagues, added check for teams on bye week.
 
 ## [1.1.0]
 
@@ -20,7 +38,15 @@ All notable changes to this project will be documented in this file.
 
 ### Initial Release
 
-[Unreleased]: https://github.com/joeyagreco/leeger/compare/v1.1.1...HEAD
+[Unreleased]: https://github.com/joeyagreco/leeger/compare/v1.3.2...HEAD
+
+[1.3.2]: https://github.com/joeyagreco/leeger/releases/tag/v1.3.2
+
+[1.3.1]: https://github.com/joeyagreco/leeger/releases/tag/v1.3.1
+
+[1.3.0]: https://github.com/joeyagreco/leeger/releases/tag/v1.3.0
+
+[1.2.0]: https://github.com/joeyagreco/leeger/releases/tag/v1.2.0
 
 [1.1.1]: https://github.com/joeyagreco/leeger/releases/tag/v1.1.1